requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
aiohttp
requests==2.27.1
beautifulsoup4==4.11.1
